Move session tuple into session.py
