+++ orchestrator/core/search/retrieval/retrievers/semantic.py
@@ -18,9 +18,10 @@ from sqlalchemy.sql.expression import ColumnElement
 
 from orchestrator.core.db.models import AiSearchIndex
 from orchestrator.core.search.core.types import EntityType, SearchMetadata
+from orchestrator.core.search.retrieval.session import HNSW_ITERATIVE_SCAN, SessionSetting
 
 from ..pagination import PageCursor
-from .base import HNSW_ITERATIVE_SCAN, Retriever, SessionSetting
+from .base import Retriever
 
 
 class SemanticRetriever(Retriever):
